Move parser and segmenter into webvtt directory
This prevents the parer library from having a naming collision with
other top level parser libraries by moving it into the webvtt directory.
