+++ lib/webvtt/parser.rb
