+++ lib/webvtt/segmenter.rb
